requirements.txt
alembic==1.14.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
dnspython==2.6.0
email_validator==2.2.0
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==21.2.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
mysqlclient==2.2.0
packaging==24.2
pkg_resources==0.0.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.21.0
